% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g}
\alias{g}
\title{Calculate the Hedge's g statistic}
\usage{
g(ordinal, interval, correct = TRUE)
}
\arguments{
\item{ordinal}{The (binary) norminal or ordinal variable.}

\item{interval}{The numberical (interval) variable.}

\item{correct}{Wheter to apply correction for chance.}
}
\description{
The function calculate the Hedge's g statistics, which is the effect size measure.
}
\examples{
g(ordinal = mtcars$vs, interval = mtcars$gear, correct = TRUE)
}
\author{
Marjan Cugmas
}
