% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marjan_color.R
\name{marjan_color}
\alias{marjan_color}
\title{Retrieve Marjan color(s)}
\usage{
marjan_color(...)
}
\arguments{
\item{...}{The names of the colors: \code{'myBlue'}, \code{'myJungle'}, \code{'myGreen'}, \code{'myRed'}, and \code{'myOrange'}. If not specified, the whole palette will be returned.}
}
\value{
A named character vector representing the color(s) from the Marjan color palette. If no specific color was requested, the entire color palette is returned.
}
\description{
This function returns hexadecimal color codes from the Marjan color palette.
}
\examples{
\dontrun{
marjan_color("myBlue", "myRed") # Returns the hexadecimal codes for "myBlue" and "myRed"
marjan_color() # Returns the entire Marjan color palette
}

}
