% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marjan_pal.R
\name{marjan_pal}
\alias{marjan_pal}
\title{Marjan Color Palette}
\usage{
marjan_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{A string representing the color palette to use. Available palettes are \code{"main"}, \code{"myBlue"}, \code{"myRed"}, \code{"myOrange"}, \code{"myGreen"}, \code{"myJungle"}, \code{"divergentPolar"}, \code{"divergentNeutral"}, \code{"divergentNeutral2"}. Default is \code{"main"}.}

\item{reverse}{Logical. If TRUE, the order of the colors in the palette is reversed. Default is FALSE.}

\item{...}{Additional arguments passed to the colorRampPalette function.}
}
\value{
A colorRampPalette function representing the chosen Marjan color palette.
}
\description{
This function returns a color palette from the Marjan color schemes. The Marjan color schemes are a set of predefined color palettes.
}
\examples{
\dontrun{
pal <- marjan_pal("myBlue")
colors <- pal(10)  # Generate 10 colors from the "myBlue" Marjan color palette

library(ggplot)
ggplot(mtcars, aes(x = gear)) +
geom_bar(fill = marjan_color("myBlue"))

}

}
\seealso{
\code{\link[grDevices]{colorRampPalette}}
}
