% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.anova.R
\name{report.anova}
\alias{report.anova}
\title{Report the ANOVA results}
\usage{
report.anova(catVarName, numVarNames, data, omegaSq = FALSE)
}
\arguments{
\item{catVarName}{A string representing the name of the ordinal factor variable in the data frame.}

\item{numVarNames}{A vector of strings representing the names of the interval variables in the data frame.}

\item{data}{A data frame or an object of class \code{survey.design} from \code{survey} package.}

\item{omegaSq}{Logical, if \code{TRUE}, the function also calculates and reports omega squared effect sizes.
By default, \code{omegaSq = FALSE}. Not implemented for weighted data.}
}
\value{
A matrix with rows representing each specified interval variable, and columns representing:
        1) means for each level of the ordinal factor (one column per level),
        2) F-statistic,
        3) degrees of freedom 1 (df1),
        4) degrees of freedom 2 (df2),
        5) p-value,
        6) optionally, omega squared effect size.
}
\description{
This function conducts one-way ANOVAs for each of the provided interval variables, with a specified ordinal variable as a factor.
It reports group means for each level of the ordinal variable, ANOVA test statistics, and optionally, omega squared effect sizes.
}
\examples{
report.anova(catVarName = "gear", numVarNames = c("mpg", "wt"),
data = mtcars, omegaSq = TRUE)
}
\references{
Kirk, R. E. (1996). Practical significance: A concept whose time has come. Educational and psychological measurement, 56(5), 746-759.
Tunks, T. (1978). The use of omega squared in interpreting statistical significance. Bulletin of the Council for Research in Music Education, 28-34.
}
\author{
Marjan Cugmas
}
