% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.freqTab.R
\name{report.freqTab}
\alias{report.freqTab}
\title{A frequency table with missing values}
\usage{
report.freqTab(
  catVarName,
  data,
  dec = 2,
  dec.freq = 0,
  cumulative = FALSE,
  useNA = "always",
  language = "english",
  sort.cat = FALSE,
  total.row = TRUE
)
}
\arguments{
\item{catVarName}{The name of a categorical variable.}

\item{data}{A data frame or an object of class \code{survey.design} from \code{survey} package.}

\item{dec}{Number of decimal places for percentages. Default is \code{2}.}

\item{dec.freq}{Number of decimal places for frequencies. Defaults is \code{0}. Only used when an object of class \code{survey.design} is specified with the parameter \code{data}.}

\item{cumulative}{Wheter to report cumulative frequencies and percentages or not. Default is \code{FALSE}.}

\item{useNA}{Wheter to include NA values in the table. Default is \code{always} (it includes NAs even when the count of NAs is zero), other ppossible values are \code{no} (do not include NAs) and \code{ifany} (NAs are included if they are present in the data). See \code{?table} for more.}

\item{language}{The language used for displaying the statistics in the frequency table. This parameter accepts two values: \code{english} or \code{slovene}. Depending on the chosen language, all statistical terms and output will be adjusted accordingly. Default is \code{english}.}

\item{sort.cat}{A logical or character value that specifies whether and how the categories of the frequency table should be sorted. If \code{FALSE} (the default), the categories are not sorted and appear in the order they are found in the data. If \code{"freq"}, the categories are sorted by frequency in descending order (most frequent categories first).}

\item{total.row}{Logical; if `TRUE`, adds a total row at the bottom of the frequency table. Default is \code{TRUE}.}
}
\description{
The function creates a frequency table (with percentages and cumulative percentages) with valid values and with missing values.
}
\note{
This is a modified function \code{freqTab} from the package \code{multiUS}. The author of the original function is Aleš Žiberna.
}
\examples{
report.freqTab(catVarName = "gear", dec = 2, useNA = "always", data = mtcars)
}
\author{
Marjan Cugmas
}
