% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.multipleChoice.R
\name{report.multipleChoice}
\alias{report.multipleChoice}
\title{Report the distributions of several dichotomous variables}
\usage{
report.multipleChoice(
  data,
  varNames,
  event = 1,
  attachLabels = FALSE,
  dec.freq = 0,
  language = "english"
)
}
\arguments{
\item{data}{A data frame or an object of class \code{survey.design} from \code{survey} package.}

\item{varNames}{A vector of strings representing the names of the variables in the data frame.}

\item{event}{A value (numeric or character) representing the event of interest to be reported.
This event is transformed to numeric 1 in each specified variable.
Default is \code{1}.}

\item{attachLabels}{Logical, if \code{TRUE}, attempts to attach variable labels from the attributes of the data frame.
Default is \code{FALSE}.}

\item{dec.freq}{Number of decimal places for frequencies. Defaults is \code{0}. Only used when an object of class \code{survey.design} is specified with the parameter \code{data}.}

\item{language}{The language used for displaying the statistics in the frequency table. This parameter accepts two values: \code{english} or \code{slovene}. Depending on the chosen language, all statistical terms and output will be adjusted accordingly. Default is \code{english}.}
}
\value{
A data frame with columns representing:
        1) the frequency of the event,
        2) the percentage of the event,
        3) the number of valid (non-NA) entries,
        4) the number of invalid (NA) entries for each specified variable.
}
\description{
This function reports the frequencies of a specified event in multiple variables of a dataset.
It converts the event of interest to numeric, calculates the frequency of the event,
the number of valid and invalid entries for each variable, and optionally attaches variable labels.
}
\examples{
data <- data.frame(A = c(1, 1, 2, NA), B = c(1, 2, 2, 1), C = c(NA, 1, 1, 1))
varNames <- c("A", "B", "C")
report.multipleChoice(data, varNames)
}
\author{
Marjan Cugmas
}
