% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.ttest.R
\name{report.ttest}
\alias{report.ttest}
\title{T-test Reporting Function}
\usage{
report.ttest(catVarName, numVarNames, data, hedgesG = FALSE, CI = FALSE)
}
\arguments{
\item{catVarName}{A string. The name of the categorical variable to be used in the t-tests.}

\item{numVarNames}{A vector of strings. The names of the numerical variables to be used in the t-tests.}

\item{data}{A data frame. The data to be used in the t-tests.}

\item{hedgesG}{A logical. If \code{TRUE}, Hedges' G effect size is computed for each t-test. Defaults to \code{FALSE}.}

\item{CI}{A logical. If \code{TRUE}, confidence intervals are computed for each t-test. Defaults to \code{FALSE}.}
}
\value{
A matrix with the results of the t-tests. Each row corresponds to one of the numerical variables,
        and the columns show the means for each level of the categorical variable, the t-statistic, the degrees of freedom,
        the p-value, and optionally the confidence intervals and Hedges' G effect size.
}
\description{
This function performs a t-test on a given categorical variable and a set of numerical variables. It also optionally computes Hedges' G effect size and confidence intervals.
}
\note{
This function assumes that the categorical variable is a factor.
      If the variable is not a factor, it will be converted to a factor within the function.
}
\examples{
report.ttest(catVarName = "am", c("mpg", "disp"), mtcars, hedgesG = TRUE, CI = TRUE)

}
\seealso{
\code{\link{t.test}}
}
\author{
Marjan Cugmas
}
