% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_marjan.R
\name{theme_marjan}
\alias{theme_marjan}
\title{Marjan's ggplot2 theme}
\usage{
theme_marjan(legend.position = "top")
}
\arguments{
\item{legend.position}{Character specifying the position of the legend on the plot.
The options are \code{"top"}, \code{"bottom"}, \code{"left"}, \code{"right"}, \code{"none"}. Default is \code{"top"}.}
}
\value{
A ggplot2 theme.
}
\description{
This function creates a custom ggplot2 theme, with various formatting
adjustments. It sets the panel grid minor and axis ticks to blank, modifies text
elements (including strip text, legend, axis titles, plot title, subtitle, and caption),
and ensures that Google font "Lato" is used where specified. The function also allows
to change the position of the legend.
}
\examples{
\dontrun{
library(ggplot2)
p <- ggplot(mtcars, aes(x = mpg, y = disp)) +
 geom_point() +
 labs(title = "Title",
 subtitle = "Subtitle",
 caption = "Caption",
 y = "Y axis title",
 x = "X axis title") +
 theme_marjan(legend.position = "bottom")
print(p)
}

}
