% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.multipleChoice.R
\name{report.multipleChoice}
\alias{report.multipleChoice}
\title{Report the distributions of several dichotomous variables}
\usage{
report.multipleChoice(data, varNames, event = 1, attachLabels = FALSE)
}
\arguments{
\item{data}{A data frame containing the variables to be analysed.}

\item{varNames}{A vector of strings representing the names of the variables in the data frame.}

\item{event}{A value (numeric or character) representing the event of interest to be reported.
This event is transformed to numeric 1 in each specified variable.
By default, event = 1.}

\item{attachLabels}{Logical, if \code{TRUE}, attempts to attach variable labels from the attributes of the data frame.
By default, \code{attachLabels = FALSE}.}
}
\value{
A data frame with columns representing:
        1) the frequency of the event,
        2) the percentage of the event,
        3) the number of valid (non-NA) entries,
        4) the number of invalid (NA) entries for each specified variable.
}
\description{
This function reports the frequencies of a specified event in multiple variables of a dataset.
It converts the event of interest to numeric, calculates the frequency of the event,
the number of valid and invalid entries for each variable, and optionally attaches variable labels.
}
\examples{
data <- data.frame(A = c(1, 1, 2, NA), B = c(1, 2, 2, 1), C = c(NA, 1, 1, 1))
varNames <- c("A", "B", "C")
report.multipleChoice(data, varNames)
}
\author{
Marjan Cugmas
}
