% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.ordinal.R
\name{report.ordinal}
\alias{report.ordinal}
\title{Report the distributions of ordinal variables}
\usage{
report.ordinal(
  ordVarNames,
  data,
  type = 1,
  dec = 1,
  dec.freq = 0,
  language = "english",
  attachLabels = FALSE
)
}
\arguments{
\item{ordVarNames}{A vector of strings representing the names of the ordinal variables in the dataset.}

\item{data}{A data frame or an object of class \code{survey.design} from \code{survey} package.}

\item{type}{An integer indicating the type of the report:
1 = proportions (default),
2 = counts,
3 = counts and proportions.}

\item{dec}{The number of decimal places to round the values.}

\item{dec.freq}{Number of decimal places for frequencies. Defaults is \code{0}. Only used when an object of class \code{survey.design} is specified with the parameter \code{data}.}

\item{language}{The language used for displaying the statistics in the frequency table. This parameter accepts two values: \code{english} or \code{slovene}. Depending on the chosen language, all statistical terms and output will be adjusted accordingly. Default is \code{english}.}

\item{attachLabels}{Logical, if \code{TRUE}, attempts to attach variable labels from the attributes of the data frame.
By default, \code{attachLabels = FALSE}.}
}
\value{
A matrix with rows representing each specified ordinal variable, and columns representing:
        1) proportions or counts (or both) of each level (one column per level),
        2) total count of valid (non-NA) entries,
        3) mean,
        4) standard deviation.
}
\description{
This function reports statistics for ordinal variables in a dataset.
It calculates proportions or counts of each level, the total count of valid (non-NA) entries, mean, and standard deviation.
}
\examples{
data <- data.frame(A = factor(c(1, 2, 3, 2), ordered = TRUE),
                   B = factor(c(2, 3, 2, 1), ordered = TRUE))
ordVarNames <- c("A", "B")
report.ordinal(ordVarNames, data, type = 1)
}
\author{
Marjan Cugmas
}
