% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.propTable.R
\name{report.propTable}
\alias{report.propTable}
\title{Create a contingency table}
\usage{
report.propTable(x, y, margin = NULL, digits = 1)
}
\arguments{
\item{x}{First categorical variable.}

\item{y}{Second categorical variable.}

\item{margin}{Index to generate margin for. If \code{NULL}, then simple probabilities (non-conditional) are reported.}
}
\description{
The function print the contingency table with columns sums or rowsums.
}
\examples{
report.propTable(x = mtcars$gear, y = mtcars$carb, margin = 1, digits = 2)
}
\author{
Marjan Cugmas
}
