% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.ttest.R
\name{report.ttest}
\alias{report.ttest}
\title{CReport the results of t-test}
\usage{
report.ttest(ordinalna, intervalna, dataset, hedgesG = FALSE)
}
\arguments{
\item{ordinalna}{The name of categorical variable. One variable can be considered.}

\item{intervalna}{The name of interval/numberic variable(s). A string or a list. Several numeric variables can be considered.}

\item{dataset}{The dataframe with categorical and numerical variables.}

\item{hedgesG}{Should Hedges' g statistics be computed?}
}
\description{
The function reports the results of t-tests with one categorical variable and one or several numeric variables.
}
\note{
Hedges' g statistics is a variation of Cohen's d, corrected for bias due to the sammple size.
}
\examples{
report.ttest(ordinalna = "vs", intervalna = c("disp", "wt", "qsec"), dataset = mtcars)
}
\references{
Hedges, L., & Olkin, I. (2014). Statistical methods for meta-analysis. Academic press.
}
\author{
Marjan Cugmas
}
